refactor: updated the structure and moved the test folder. Added few improvements to the config files. Removed unnecessary tests.
