+++ backend/eslint.config.mjs
@@ -2,7 +2,7 @@ import globals from "globals";
 import tseslint from "typescript-eslint";
 import eslintPluginPrettierRecommended from "eslint-plugin-prettier/recommended";
 
-export default tseslint.config(
+export default tseslint.config([
   // Block 1: Global Ignores
   {
     ignores: ["dist/", "node_modules/"],
@@ -39,5 +39,5 @@ export default tseslint.config(
 
   // Prettier Integration (must be last)
   eslintPluginPrettierRecommended,
-);
+]);
