+++ backend/tsconfig.json
@@ -4,9 +4,7 @@
     "outDir": "./dist",
     "rootDir": "./src",
     "types": ["node"],
-    "esModuleInterop": true,
     "resolveJsonModule": true,
-    "module": "CommonJS",
   },
   "files": ["src/types/fastify.d.ts"],
   "include": ["src/**/*.ts"],
